Apply asset loading to history chrome. Also added a `globals.history.items_skelly` and `globals.history.search_select` for more configurability of history items.
